pyproject.toml
[project]
name = "pr-lint"
version = "0.1.0"
description = ""
authors = [{ name = "lucemia", email = "lucemia@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pygithub",
    "typer",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "syrupy",
    "pre-commit>=4.0.1",
    "ruff>=0.8.0",
]

[project.scripts]
pr-lint = "pr_lint.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pr_lint"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build]
include = ["pr_lint/py.typed"]
exclude = ["**/tests"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "B905"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
